Remove obsolete comment (#20)
This comment is well over a year old, so I'm removing it.  I see a `.travis` file in this repo and a `Podfile` in the example; but I don't speak Xcode, so I have no idea which things mentioned in the comment are still needed.  If there's any cleanup work that still needs to happen, please open an issue for it. :smiley: